Add files via upload
This has javadoc comments for FrontEnd and BackEnd. I was not sure if we should do comments for the other classes or not because I did not code them.
I looked into implementing the admin/ user database stuff for bonus marks but it looked like quite a bit of work and changes to the program (which is already not super stable lol). I hope that is okay with you, I had meant to get it done on Saturday but fixing the bugs and getting the gui working with the basic requirements took a lot longer than expected.